+++ src/main/resources/META-INF/services/me.fengming.openjs.plugin.OpenJSPluginLoader
@@ -1,1 +1,1 @@
-me.fengming.openjs.plugin.OpenJSBuiltinPluginLoader
+me.fengming.openjs.plugin.builtin.OpenJSBuiltinPluginLoader
